improves project structure
