+++ src/actions/sendMessage.ts
@@ -2,10 +2,13 @@ import axios from "axios";
 
 import "dotenv/config";
 
-import { getChannelIdRealtime, runtimeChannelData } from "../utils";
+import { getChannelIdRealtime, runtimeChannelData } from "../utils/utils";
 
-export const sendMessage = async (channel: string, messageContent: string) => {
-  const channelId = getChannelIdRealtime(runtimeChannelData, channel);
+export const sendMessage = async (
+  channelUsername: string,
+  messageContent: string
+) => {
+  const channelId = getChannelIdRealtime(runtimeChannelData, channelUsername);
   try {
     const axiosRequest = await axios.post(
       `https://kick.com/api/v2/messages/send/${channelId}`,
@@ -20,18 +23,18 @@ export const sendMessage = async (channel: string, messageContent: string) => {
           "content-type": "application/json",
           "x-xsrf-token": process.env.TOKEN,
           cookie: process.env.COOKIES,
-          Referer: `https://kick.com/${channel}`,
+          Referer: `https://kick.com/${channelUsername}`,
         },
       }
     );
 
     if (axiosRequest.status === 200) {
       console.log(
-        `✅ SUCCESFULLY SENT A REPLY BACK - ${channelId}: ${messageContent}`
+        `✅ SUCCESFULLY SENT A MESSAGE BACK - ${channelId}: ${messageContent}`
       );
     } else {
       console.log(
-        `💥 ERROR SENDING A REPLY - STATUS CODE: ${axiosRequest.status}; ${axiosRequest.statusText} EXTRA DETAILS: CHANNEL: ${channelId}; MESSAGE: ${messageContent}`
+        `💥 ERROR SENDING A MESSAGE - STATUS CODE: ${axiosRequest.status}; ${axiosRequest.statusText} EXTRA DETAILS: CHANNEL: ${channelId}; MESSAGE: ${messageContent}`
       );
     }
   } catch (error) {
