+++ src/commands/test.ts
@@ -1,4 +1,4 @@
-import { sendMessage } from "../handlers/sendMessage";
+import { sendMessage } from "../actions/sendMessage";
 
 export const execute = (command: string, username: string, channel: string) => {
   const randomNumber = Math.floor(Math.random() * 999999) + 1;
